refactor iocage.lib.Config.Properties
